backend/requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
google-auth==2.23.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
httpx-oauth==0.13.0
idna==3.4
Jinja2==3.1.2
kubernetes==28.1.0
makefun==1.15.1
MarkupSafe==2.1.3
oauthlib==3.2.2
passlib==1.7.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.4.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.18
uvicorn==0.23.1
websocket-client==1.6.3
